requirements.txt
# Core LLM + LangChain
openai>=1.30.1          # GPT-4o & structured-output support
langchain-openai>=0.1.2 # LangChain wrapper for OpenAI
langgraph>=0.0.41       # Graph execution engine used in the agent
langgraph-checkpoint-sqlite

# Retrieval
tavily-python>=0.3.2    # Web search API client
tabulate

# Data wrangling
pandas>=2.2.2
reportlab
# Web UI
streamlit>=1.33.0

# Utilities
python-dotenv>=1.0.1    # Load .env files
